+++ packages/icons/src/svg/index.native.js
@@ -1,3 +1,6 @@
 import Star from './star.svg';
+import Clock from './clock.svg';
+import Building from './building.svg';
+import TriangleAlert from './triangle_alert.svg';
 
-export { Star };
+export { Star, Clock, Building, TriangleAlert };