+++ packages/yoga/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.9.0](https://github.com/Gympass/yoga/compare/@gympass/yoga@1.8.0...@gympass/yoga@1.9.0) (2020-01-21)
+
+### Features
+
+- **tag:** new component: Tag ([7bbb45c](https://github.com/Gympass/yoga/commit/7bbb45ce282e562d97f3a3134fa8db01b6600e8e))
+- **tag:** new component: Tag (web and native) ([659af81](https://github.com/Gympass/yoga/commit/659af8155d60bec4f540a90b4bf7c9a76512fbee))
+
+# [1.8.0](https://github.com/Gympass/yoga/compare/@gympass/yoga@1.7.0...@gympass/yoga@1.8.0) (2020-01-20)
+
+### Features
+
+- **gymcard.checkin:** create new GymCard.CheckIn component ([c1d0d83](https://github.com/Gympass/yoga/commit/c1d0d83d55984d4fb99fdd8b9bc36df90bf31d54))
+
+# [1.7.0](https://github.com/Gympass/yoga/compare/@gympass/yoga@1.6.2...@gympass/yoga@1.7.0) (2020-01-20)
+
+### Bug Fixes
+
+- **card:** adjusting variant property on event card ([bfeedee](https://github.com/Gympass/yoga/commit/bfeedee863ca5f2bbc37ab2e6b1a0d4303bb7170))
+
+### Features
+
+- **card:** adding clock icon in the event card ([98dac45](https://github.com/Gympass/yoga/commit/98dac45988d05d70a789fe1afffee7598f2f6ebd))
+- **card:** creating card event ([1a40aa1](https://github.com/Gympass/yoga/commit/1a40aa16357ff30d6133feaed4048c679f5682c6))
+- **card:** creating event card for web ([ea9ed3c](https://github.com/Gympass/yoga/commit/ea9ed3cfe294f9b00b0c8204753651a265857c8b))
+
 ## [1.6.2](https://github.com/Gympass/yoga/compare/@gympass/yoga@1.6.1...@gympass/yoga@1.6.2) (2020-01-16)
 
 ### Bug Fixes